@@ -1,12 +1,12 @@
 apply plugin: 'com.android.library'
 
 android {
-    compileSdkVersion 23
-    buildToolsVersion '25.0.0'
+    compileSdkVersion 28
+    buildToolsVersion '28.0.3'
 
     defaultConfig {
         minSdkVersion 16
-        targetSdkVersion 23
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
 
@@ -25,11 +25,16 @@ android {
 }
 
 dependencies {
-    compile 'com.facebook.react:react-native:+'
-    compile 'com.google.code.gson:gson:2.8.0'
-    testCompile 'junit:junit:4.12'
-    compile fileTree(include: ['*.jar'], dir: 'libs')
-    compile files('libs/jpush-android_v3.0.6.jar')
-    compile 'com.squareup.picasso:picasso:2.5.2'
-    compile "me.leolin:ShortcutBadger:1.1.16@aar"
+    implementation 'com.facebook.react:react-native:+'
+    implementation 'com.google.code.gson:gson:2.8.0'
+    testImplementation 'junit:junit:4.12'
+    implementation fileTree(include: ['*.jar'], dir: 'libs')
+    implementation files('libs/jpush-android_v3.0.6.jar')
+    implementation 'com.squareup.picasso:picasso:2.5.2'
+    implementation "me.leolin:ShortcutBadger:1.1.16@aar"
+    //compileOnly 'androidx.appcompat:appcompat:1.2.0'
+
+    implementation 'androidx.appcompat:appcompat:1.2.0'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.1'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
 }
