@@ -5,12 +5,13 @@ import android.app.Dialog;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
 import android.util.DisplayMetrics;
 import android.util.Log;
 import android.view.View;
 import android.view.WindowManager;
 
+
+import androidx.fragment.app.Fragment;
 
 import com.xsdlr.rnjmessage.R;
 import com.xsdlr.rnjmessage.im.chatting.utils.DialogCreator;
