@@ -14,8 +14,9 @@ import android.graphics.BitmapFactory;
 import android.graphics.Point;
 import android.os.Handler;
 import android.os.Message;
-import android.support.v4.util.LruCache;
 
+
+import androidx.collection.LruCache;
 
 import com.xsdlr.rnjmessage.im.chatting.utils.BitmapLoader;
 
