@@ -1,9 +1,10 @@
 package com.xsdlr.rnjmessage.im.view;
 
 import android.content.Context;
-import android.support.v4.view.ViewPager;
 import android.util.AttributeSet;
 import android.view.MotionEvent;
+
+import androidx.viewpager.widget.ViewPager;
 
 
 public class ScrollControlViewPager extends ViewPager {
